+++ Command/InvokableCommand.php
@@ -33,7 +33,6 @@ class InvokableCommand implements SignalableCommandInterface
     private readonly \Closure $code;
     private readonly ?SignalableCommandInterface $signalableCommand;
     private readonly \ReflectionFunction $reflection;
-    private bool $triggerDeprecations = false;
 
     public function __construct(
         private readonly Command $command,
@@ -52,12 +51,6 @@ class InvokableCommand implements SignalableCommandInterface
         $statusCode = ($this->code)(...$this->getParameters($input, $output));
 
         if (!\is_int($statusCode)) {
-            if ($this->triggerDeprecations) {
-                trigger_deprecation('symfony/console', '7.3', \sprintf('Returning a non-integer value from the command "%s" is deprecated and will throw an exception in Symfony 8.0.', $this->command->getName()));
-
-                return 0;
-            }
-
             throw new \TypeError(\sprintf('The command "%s" must return an integer value in the "%s" method, but "%s" was returned.', $this->command->getName(), $this->reflection->getName(), get_debug_type($statusCode)));
         }
 
@@ -86,8 +79,6 @@ class InvokableCommand implements SignalableCommandInterface
         if (!$code instanceof \Closure) {
             return $code(...);
         }
-
-        $this->triggerDeprecations = true;
 
         if (null !== (new \ReflectionFunction($code))->getClosureThis()) {
             return $code;
@@ -124,12 +115,6 @@ class InvokableCommand implements SignalableCommandInterface
             $type = $parameter->getType();
 
             if (!$type instanceof \ReflectionNamedType) {
-                if ($this->triggerDeprecations) {
-                    trigger_deprecation('symfony/console', '7.3', \sprintf('Omitting the type declaration for the parameter "$%s" is deprecated and will throw an exception in Symfony 8.0.', $parameter->getName()));
-
-                    continue;
-                }
-
                 throw new LogicException(\sprintf('The parameter "$%s" must have a named type. Untyped, Union or Intersection types are not supported.', $parameter->getName()));
             }